## Updates - Split the search screen into two separate screens. - Updated the navigation and routes to be more consistent. - Refactored naming, packages and structure.
