+++ lib/pages/full_music_links/full_music_links_screen.dart
@@ -3,21 +3,26 @@ import 'package:syncrosong/utility/widgets/loader_widget.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 
-class MusicLinksWebView extends StatefulWidget {
-  final WebViewController webViewController;
+class FullMusicLinksScreen extends StatefulWidget {
+  final WebViewController webViewController = WebViewController();
+  final String url;
 
-  const MusicLinksWebView(this.webViewController, {super.key});
+  FullMusicLinksScreen(String? url, {super.key}) : url = url ?? "";
 
   @override
-  State<MusicLinksWebView> createState() => _MusicLinksWebViewState();
+  State<FullMusicLinksScreen> createState() => _FullMusicLinksScreenState();
 }
 
-class _MusicLinksWebViewState extends State<MusicLinksWebView> {
-  bool isLoading = true;
+class _FullMusicLinksScreenState extends State<FullMusicLinksScreen> {
+  bool _isLoading = true;
 
   @override
   void initState() {
     super.initState();
+
+    Uri uri = Uri.parse(widget.url);
+    uri.replace(scheme: "https");
+    widget.webViewController.loadRequest(uri);
 
     widget.webViewController.setJavaScriptMode(JavaScriptMode.unrestricted);
     widget.webViewController.setNavigationDelegate(NavigationDelegate(
@@ -29,30 +34,24 @@ class _MusicLinksWebViewState extends State<MusicLinksWebView> {
 
   @override
   Widget build(BuildContext context) {
-    if (isLoading) {
-      return Container(
-          decoration: const BoxDecoration(color: Colors.white),
-          constraints: const BoxConstraints.expand(width: double.infinity, height: double.infinity),
-          child: const LoaderWidget());
-    } else {
-      return SingleChildScrollView(
-        child: ConstrainedBox(
-          constraints: BoxConstraints(maxHeight: MediaQuery
-              .of(context)
-              .size
-              .height),
-          child: Column(mainAxisSize: MainAxisSize.min, children: [
-            SizedBox(height: MediaQuery
-                .of(context)
-                .viewPadding
-                .top),
-            Flexible(
-              child: WebViewWidget(controller: widget.webViewController),
+    return Scaffold(
+      body: _isLoading
+          ? Container(
+              decoration: const BoxDecoration(color: Colors.white),
+              constraints: const BoxConstraints.expand(width: double.infinity, height: double.infinity),
+              child: const LoaderWidget())
+          : SingleChildScrollView(
+              child: ConstrainedBox(
+                constraints: BoxConstraints(maxHeight: MediaQuery.of(context).size.height),
+                child: Column(mainAxisSize: MainAxisSize.min, children: [
+                  SizedBox(height: MediaQuery.of(context).viewPadding.top),
+                  Flexible(
+                    child: WebViewWidget(controller: widget.webViewController),
+                  ),
+                ]),
+              ),
             ),
-          ]),
-        ),
-      );
-    }
+    );
   }
 
   void _onPageFinishedLoading() {
@@ -63,7 +62,7 @@ class _MusicLinksWebViewState extends State<MusicLinksWebView> {
           document.getElementsByClassName('css-d8rran')[0].style.display='none';
         })()""");
     setState(() {
-      isLoading = false;
+      _isLoading = false;
     });
   }
